Dutch National Flag problem in Java
